Issue 1: IndexError on Plotting The IndexError seems to be caused by an attempt to access a list index that does not exist. This could be due to an empty DataFrame or incorrect handling of the data. We need to ensure that the data we are working with is correctly filtered and non-empty.
Issue 2: Dynamic Y-Axis Labeling
To dynamically change the Y-axis label based on the data, we can implement a function that determines the appropriate unit (ms, s, min) based on the maximum value in the data.

Issue 3: Sorting High Latency Periods
Sorting the dropdown and table can be achieved by sorting the DataFrame before displaying it.